Update to use v2 api endpoint.
The v1 endpoint will be completely disable in June 2018. This commit
moves to the v2 api.
The primary differences are authentication is handled by the
Authorization HTTP header and processing of requests is deferred.

As such a requests return a 202 status to indicate they've been queued
for processing, and return a requestId in the body.
This requestId can then be used to poll the status of the request to
confirm if it was successful.
Right now we don't do this, but could look at adding it in the future.