Close matching open alerts on job success if enabled
Using the Alias field, with a sensible default value of ${job.id}
a job can be configured to automatically close an open alert on
success.

The Rundeck notification settings are a little unclear , but just enable the
notification on job success, and ensure the Alias matches that used in
the error notification.